LPS-57210 Move portlet keys to API bundle
