+++ modules/apps/mobile-device-rules/mobile-device-rules-api/src/com/liferay/mobile/device/rules/constants/MobileDeviceRulesPortletKeys.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.mobile.device.rules.web.constants;
+package com.liferay.mobile.device.rules.constants;
 
 /**
  * @author Mate Thurzo
